Separate COMM & RTC panels
